Add priority functionality
The priority will decide if a button have priority to write to the
output.
Priority 0 highest
Priority 255 lowest

It takes parameters .CTRLParam0 and .CTRLParam1 to set up the priorities
to the relays and status outputs.

When a input is setup as SetPriority, when the input is asserted
it will assign the priority stored in .CTRLParam0 to the corresponding
relays (.CTRLTargetRelay) and status output (.CTRLTargetStatus).

When the input is de-asserted it will assign the priority stored
in .CTRLParam1 to the corresponding relays (.CTRLTargetRelay) and
status output (.CTRLTargetStatus).